pyproject.toml
[tool.poetry]
name = "esales1"
version = "0.1.0"
description = ""
authors = ["Evlit <evlit23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
envparse = "^0.2.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
